use shared component for removing species confirmation message
