+++ src/content/app/species/components/species-title-area/SpeciesTitleArea.tsx
@@ -14,27 +14,36 @@
  * limitations under the License.
  */
 
-import React, { useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
 import classNames from 'classnames';
+import { useNavigate } from 'react-router-dom';
+
+import * as urlFor from 'src/shared/helpers/urlHelper';
 
 import { useAppSelector, useAppDispatch } from 'src/store';
+import useSpeciesAnalytics from 'src/content/app/species/hooks/useSpeciesAnalytics';
 
 import { getDisplayName } from 'src/shared/components/selected-species/selectedSpeciesHelpers';
+import { SecondaryButton } from 'src/shared/components/button/Button';
+import DeletionConfirmation from 'src/shared/components/deletion-confirmation/DeletionConfirmation';
 import SearchIcon from 'static/icons/icon_search.svg';
 
 import { isSpeciesSidebarOpen as getSidebarStatus } from 'src/content/app/species/state/sidebar/speciesSidebarSelectors';
 import { getCommittedSpeciesById } from 'src/content/app/species-selector/state/speciesSelectorSelectors';
 import { getActiveGenomeId } from 'src/content/app/species/state/general/speciesGeneralSelectors';
 import { getPopularSpecies } from 'src/content/app/species-selector/state/speciesSelectorSelectors';
+
 import {
   SpeciesSidebarModalView,
   updateSpeciesSidebarModalForGenome
 } from 'src/content/app/species/state/sidebar/speciesSidebarSlice';
 
-import { fetchPopularSpecies } from 'src/content/app/species-selector/state/speciesSelectorSlice';
+import {
+  fetchPopularSpecies,
+  deleteSpeciesAndSave
+} from 'src/content/app/species-selector/state/speciesSelectorSlice';
 
 import SpeciesUsageToggle from './species-usage-toggle/SpeciesUsageToggle';
-import SpeciesRemove from './species-remove/SpeciesRemove';
 
 import { RootState } from 'src/store';
 
@@ -70,7 +79,10 @@ const SpeciesTitleArea = () => {
   const activeGenomeId = useAppSelector(getActiveGenomeId);
   const isSidebarOpen = useAppSelector(getSidebarStatus);
   const dispatch = useAppDispatch();
+  const navigate = useNavigate();
   const { species, iconUrl } = useSpecies() || {};
+  const [isRemoving, setIsRemoving] = useState(false);
+  const { trackDeletedSpecies } = useSpeciesAnalytics();
 
   const blockClasses = classNames(styles.speciesTitleArea, {
     [styles.speciesTitleAreaNarrow]: isSidebarOpen
@@ -87,6 +99,16 @@ const SpeciesTitleArea = () => {
         fragment: { sidebarModalView: SpeciesSidebarModalView.SEARCH }
       })
     );
+  };
+
+  const toggleRemovalDialog = () => {
+    setIsRemoving(!isRemoving);
+  };
+
+  const onRemove = () => {
+    dispatch(deleteSpeciesAndSave(activeGenomeId));
+    species && trackDeletedSpecies(species);
+    navigate(urlFor.speciesSelector());
   };
 
   return species && iconUrl ? (
@@ -106,8 +128,24 @@ const SpeciesTitleArea = () => {
         <SearchIcon />
       </div>
       <div className={styles.speciesRemove}>
-        <SpeciesRemove />
+        <SecondaryButton
+          onClick={toggleRemovalDialog}
+          isDisabled={isRemoving}
+          className={classNames({ [styles.disabledRemoveButton]: isRemoving })}
+        >
+          Remove species
+        </SecondaryButton>
       </div>
+      {isRemoving && (
+        <DeletionConfirmation
+          warningText="If you remove this species, any views you have configured will be lost — do you wish to continue?"
+          confirmText="Remove"
+          cancelText="Do not remove"
+          className={styles.speciesRemoveMessage}
+          onCancel={toggleRemovalDialog}
+          onConfirm={onRemove}
+        />
+      )}
     </div>
   ) : (
     <div className={blockClasses} />
