--- src/content/app/species/components/species-title-area/species-remove/SpeciesRemove.scss
@@ -1,47 +1,0 @@
-@import 'src/styles/common';
-
-// these two variables are used to emulate the unavailable flexbox-gap property
-$fake-top-margin: 6px;
-$fake-left-margin: 60px;
-
-.speciesRemovalConfirmation {
-  /*
-    FIXME when Safari implements gap for flexbox
-    or when it becomes possible to test support for flexbox gap.
-    Will be something as simple as:
-
-    gap: 25px;
-  */
-  margin-top: -#{$fake-top-margin};
-  margin-left: -#{$fake-left-margin};
-
-  & > * {
-    margin-top: $fake-top-margin;
-    margin-left: $fake-left-margin;
-  }
-}
-
-.speciesRemovalWarning {
-  display: inline-block;
-  color: $red;
-  line-height: 1.2;
-}
-
-.speciesRemovalConfirmationControls {
-  margin-left: $fake-left-margin;
-  display: inline-block;
-
-  button {
-    margin-right: 35px;
-  }
-}
-
-.clickable {
-  color: $blue;
-  cursor: pointer;
-}
-
-.buttonContainer {
-  display: flex;
-  justify-content: flex-end;
-}
